refactor: update home components imports to use shadcn directory
