+++ features/home/components/marketing-footer.tsx
@@ -3,8 +3,8 @@
 import Image from 'next/image';
 import Link from 'next/link';
 import { ArrowUpRight } from 'lucide-react';
-import { Button } from '@/components/ui/button';
-import { Badge } from '@/components/ui/badge';
+import { Button } from '@/components/shadcn/button';
+import { Badge } from '@/components/shadcn/badge';
 
 export function MarketingFooter() {
   const year = new Date().getFullYear();
