+++ features/home/components/trusted-by-section.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { Badge } from '@/components/ui/badge';
+import { Badge } from '@/components/shadcn/badge';
 
 export function TrustedBySection() {
   const companies = [
